% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_measure.R
\name{evaluate_class}
\alias{evaluate_class}
\alias{calculate_confusion}
\title{Evaluate Class}
\usage{
evaluate_class(actual, predicted, metric = "f1")

calculate_confusion(actual, predicted)
}
\arguments{
\item{actual}{\code{numeric} \cr
initial cluster assignments}

\item{predicted}{\code{numeric} \cr
cluster assignments of permuted data}

\item{metric}{\code{character(1)} \cr
binary score metric}
}
\value{
A binary score for each of the clusters and the number of instances.
}
\description{
Calculation of binary similarity metric based on confusion matrix.
}
\section{Functions}{
\itemize{
\item \code{calculate_confusion()}: Calculate confusion matrix

}}
