% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pare_tau}
\alias{pare_tau}
\title{Pare Tau}
\arguments{
\item{R;}{numeric matrix of correlations.}

\item{tau;}{numeric vector of \eqn{\tau} values.}
}
\value{
\verb{numeric vector} of original \eqn{\tau} values with
redundant elements removed.
}
\description{
Removes redundant \eqn{\tau} values given a correlation matrix
\eqn{R}.
}
\details{
Two \eqn{\tau} values are considered redundant if they result in
the same thresholded graph. Redundant \eqn{\tau} are removed prior to
clique analysis for efficient computation.
}
